Rudementary Man-In-The_Middle attack for abacus.
YOU NEED A PROPERLY SIGNED CERTIFICATE FOR THIS TO WORK!

You need ucspi-ssl (http://www.superscript.com/ucspi-ssl/intro.html) and
OpenSSL for this to work.

It doesn't seem to be particularly stable in that it will only handle a single
connection before it needs restarting, but that should be good enough for
debugging purposes.  At the moment it assumes the destination server is
listening on localhost:7368.


git-svn-id: https://abacuscm.svn.sourceforge.net/svnroot/abacuscm/trunk@418 ac3b2d11-31fc-0310-b7bc-aecf6fe4fc4f