Add a thread to check for marker timeouts.
Every time a submission is issued for marking, we note the time that this happened. The timeout checking thread wakes up once per second, and sees if any of the issued submissions have taken too long to mark. If so, the submission is de-issued and re-enqueued for marking, and the associated client connection is removed from the set of markers. This means that we'll need to be on the lookout for submissions building up, but will mean that we don't end up with a permanently wedged marker.


git-svn-id: https://abacuscm.svn.sourceforge.net/svnroot/abacuscm/branches/it_challenge@642 ac3b2d11-31fc-0310-b7bc-aecf6fe4fc4f