Major refactoring of permissions
Permissions are now stored in a PermissionSet in the client, and the old
PermissionSet is now called PermissionTest. Permissions are also sent to
the client in the response to 'whatami' (instead of the user type), and
the client enables or disables controls based entirely on the
permissions instead of the user type.

There are now very few parts of the system that know anything about user
types:
- Adding a new user shows a standard list of user types, and users are
  stored in the database with a type
- PermissionMap hard-codes the type->permission mappings
- Logging in applies the type->permission mappings to store in the
  client connection
- When determining whether PERMISSION_JUDGE is sufficient to set a mark,
  the type of the user who previously assigned a mark is examined.

git-svn-id: https://abacuscm.svn.sourceforge.net/svnroot/abacuscm/branches/it_challenge@751 ac3b2d11-31fc-0310-b7bc-aecf6fe4fc4f