Attempt 1 at converting from Qt3 => Qt4.
This is a poor attempt if I have to say so myself.

There are too many intermediate classes and crap being added that
serves no purpose other than to glue things in such a way as to
reduce code change elsewhere.

To put the cherry on the cake, it doesn't even work.  It segfaults.