Partial implementation of judge support.
The judge may now click on submissions and see the files, but cannot
assign marks. Also haven't tested error handling for things like
clicking on a pending submission. The <pre> with the data should
probably be replaced by a <textarea> so that things scroll within the
textarea rather than making the entire webapp enormous (would also be
easier to copy-and-paste).

The contestant's source is also not shown.