Fix a serious bug in receiving content
If there is more than ~16KiB of message content, it could be truncated
(due to the way BufferedInputStream works), which then causes the
following message to be gibberish (content parsed as header). This
presumably didn't show up earlier because content blobs were only being
used for compiler errors, which are seldom 16KiB large.