Add replaceproblem command to abacuscm (see #52)
This just provides the command-line interface to the existing code. It
does not yet handle re-evaluation of previously marked runs, and I don't
know what will happen to runs that are queued for marking.

There are probably also other subtleties that aren't considered,
particularly if the dependency graph is modified (for one thing, it's
now possible to create a dependency cycle).