More assorted Docker work
The main change is the addition of inichange.py, which makes replacement
of config settings easier and more robust (sed would break on passwords
containing a /). /conf now only contains override files (if
needed), and the final config file is written in /data. The marker seems
to be working, but security needs to be locked down so that user nobody
cannot access any credentials (and ideally, we'd create a virtualenv for
a Python chroot jail).