Show message when the scoreboard is frozen.
A related change is that specifying blinds = 0 will really set no
blinds, instead of setting them to 1 hour. This is useful for unfreezing
the scoreboard by restarting the server with blinds = 0.

The client-server protocol is extended to return the blind time in the
contesttime reply, and the client then decides whether it is blinded.
This isn't perfect (in particular, there is a window where some
submissions have been received in time by the server, but not yet
marked).