Fix up batch test to run against Docker-based server
Mostly this is just handling the fact that the Docker-based server is
pre-loaded from marker and standings accounts and this changes the IDs.
Eventually it may be worth re-writing the tests in a higher-level
language that can react to whatever IDs the server assigns.

There were also a few genuine bugs in the tests.