Switch from pyinstaller to cx_Freeze
This avoids the pyinstaller issue with the user code being run in a
child process. It also simplifies the build, because cx_Freeze doesn't
try to create a bunch of build files.

This is working well for Python 2, but Python 3 fails with
Fatal Python error: Failed to open /dev/urandom