Use BuildKit in Dockerfile for mvn caching
This replaces the previous approach, which was to manually run a step to
download a lot of the mvn packages that would be needed, early on in the
build so that they'd use the normal Docker build cache. With BuildKit
one can cache a directory for a particular command, which provides much
more accurate and complete caching.